+++ doc/source/changelog.rst
@@ -10,6 +10,7 @@ This version of pyXSIM contains significant new features, and some
 backwards-incompatible changes.
 
 * Charge exchange
+* Trace abundances
 * Support for Doppler shifting of intensity fields and spectra created from data
   objects has been added. See :ref:`intensity-fields` and :ref:`xray-spectra`
   for more details.